refactor(gui): decompose Flask routes into blueprints
