+++ rex/routes/__init__.py
@@ -1,0 +1,1 @@
+"""Rex route blueprints package."""
